frontend/requirements.txt
av
aiortc
aiohttp==3.8.3
google-cloud-storage==2.7.0
ipykernel==6.19.2
openmim==0.3.5
pip-chill==1.0.1
pysocks==1.7.1
streamlit-nested-layout==0.1.1
streamlit-on-hover-tabs==1.0.1
streamlit-webrtc==0.44.2
uvicorn==0.20.0